chore(package): relax Node and npm version requirements
Relax the `node` and `npm` version requirements specified in the
`engines` property of the project's `package.json` file.

Yarn checks the version range specified in a package's
`engines.node` property of the `package.json` file.

If you attempt to install a package using a Node version that does not
fall within that package's `engines.node` version range, Yarn will fail
during the install process. This is opposed to printing a warning like
the `npm` package manager does in this situation.

To install a package using a Node version outside of the package's range
you can pass the `--ignore-engines` flag to `yarn`.

Yarn issue related to the strict treatment of the `engines` data:
- https://github.com/yarnpkg/yarn/issues/1102

We are relaxing the range since it does not seem Yarn will, any time
soon, provide a more convenient means of installing packages with
stricter `node` and `npm` version ranges.

While our `node` and `npm` version range was meant to correspond with
our Node LTS support policy, the disruption is too great to our
community. Instead, we will simply need to communicate our support
policy directly when engaging our community.