fix(index): support releases from a branch
We want to test the ability to run `npm-publish-git-tag` off of a
branch. Occasionally people will encounter the following scenario:

Someone has released a new major version of their project. A consumer
of that project reports a bug in the earlier major version, and can't,
for whatever reason, upgrade to the latest major version at this time.
That consumer would greatly benefit if the project could quickly submit
a patch against the earlier major version and have
`npm-publish-git-tag` automatically publish that version.

The owner of the project should be able to create a dedicated branch
off of the latest code for the previous major version, push a bug
fix to that branch, and after having a new tag created on that branch,
have `npm-publish-git-tag` automatically publish the new patch version.