fix: return non-zero exit code when npm publish fails
Return a non-zero exit code if the `npm publish` command fails for
any reason.

BREAKING CHANGE:

Past behavior was that when `npm publish` failed, the CLI and the
programmatic interface would ignore the exit code of `npm` and simply
return a success exit code (0).

Now, when `npm` fails and returns a non-zero exit code, the CLI will
return a non-zero exit code, and the programmatic interface will throw
an error.

Fixes #9