feat(cli): pass token as CLI option
Pass token as CLI option instead of passing the token as an environment
variable.

BREAKING CHANGE:

If your target npm registry does not require authentication, or your
project's `.npmrc` configuration file is already setup properly for
publishing, then you no longer need to set `--skipToken` to bypass
the token requirement of this tool.

If you have previously set `NPM_TOKEN` environment variable before
calling this tool, then you will need to pass that environment variable
as either a CLI option, or, if you're calling this tool programmatically,
then you pass the token as an argument to the package.

```bash
npm-deploy-git-tag --token ${NPM_TOKEN}
```

This change was done to make it obvious that a token is required by
the deployment/publishing process, removing the _magic_ behavior of
pulling in the token from an environment variable.

This also eliminates the need for `--skipToken` for those that don't
require authentication for publishing.

Closes #7